services/download-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.1
python-jose[cryptography]==3.3.0
cassandra-driver==3.29.2
minio==7.2.12
